rpl: store MinHopRankIncrease as u16
WI_SUN-3765

+++ app_wsbrd/rpl/rpl.h
@@ -105,7 +105,7 @@ struct rpl_root {
     uint8_t dodag_id[16];
     uint8_t dodag_version_number;
     uint8_t dodag_pref;
-    uint8_t min_rank_hop_inc;
+    uint16_t min_rank_hop_inc;
     uint32_t lifetime_s;
     uint16_t lifetime_unit_s;
     uint8_t dtsn; // DAO Trigger Sequence Number
